reorg UI components and utilities to load like main Romo does
This makes sure the load order works out things are organized
and initialize properly. This follows the same pattern for the
UI components that the main Romo components use.
