+++ lib/ui.js
@@ -1,27 +1,3 @@
-import './ui/dropdown.js'
-import './ui/dropdown_form.js'
-import './ui/nav.js'
-import './ui/popover.js'
-import './ui/popover_form.js'
-import './ui/popover_stack.js'
-
-class RomoUI {
-  constructor() {
-    this.popoverStackSelectors = []
-  }
-
-  setup() {
-    Romo.setup()
-  }
-
-  addPopoverStackSelector(selector) {
-    this.popoverStackSelectors.push(selector)
-  }
-}
-
-Romo.ui = new RomoUI()
-
-Romo.addPopoverStackSelector =
-  function(selector) {
-    Romo.ui.addPopoverStackSelector(selector)
-  }
+import './ui/env.js'
+import './ui/utilities.js'
+import './ui/components.js'
