+++ lib/ui/components.js
@@ -1,0 +1,5 @@
+import './components/dropdown.js'
+import './components/dropdown_form.js'
+import './components/nav.js'
+import './components/popover.js'
+import './components/popover_form.js'
