+++ lib/ui/components/dropdown.js
@@ -1,4 +1,4 @@
-import '../components/dom_component.js'
+import '../../components/dom_component.js'
 import './dropdown/css_position_data.js'
 import './dropdown/placement_config.js'
 import './dropdown/placement_data.js'
