+++ lib/ui/components/dropdown/available_height_px.js
