+++ lib/ui/components/dropdown/available_width_px.js
