+++ lib/ui/components/popover_form.js
@@ -1,5 +1,5 @@
-import '../components/dom_component.js'
-import '../components/form.js'
+import '../../components/dom_component.js'
+import '../../components/form.js'
 
 Romo.define('Romo.UI.PopoverForm', function() {
   return class extends Romo.DOMComponent {
