+++ lib/ui/utilities/popover_stack.js
